change leaderboard week dates dynamically
