+++ lib/utils.js
@@ -11,25 +11,26 @@ export const humanize = (str) => {
     });
 };
 
+const months = [
+  "Jan",
+  "Feb",
+  "Mar",
+  "Apr",
+  "May",
+  "Jun",
+  "July",
+  "Aug",
+  "Sep",
+  "Oct",
+  "Nov",
+  "Dec",
+];
+
 export const parseDateString = (dateStr) => {
   const date = new Date(dateStr);
 
   if (date == "Invalid Date") return "";
 
-  const months = [
-    "Jan",
-    "Feb",
-    "Mar",
-    "Apr",
-    "May",
-    "Jun",
-    "July",
-    "Aug",
-    "Sep",
-    "Oct",
-    "Nov",
-    "Dec",
-  ];
   let dateString = date.getDate() + " ";
   dateString += months[date.getMonth()] + " ";
   dateString += date.getFullYear() + " | ";
@@ -49,3 +50,20 @@ export const parseDateString = (dateStr) => {
   dateString += date.getHours() > 11 ? " PM" : " AM";
   return dateString;
 };
+
+// Calculate week number
+export const getWeekNumber = (date) => {
+  const d = new Date(date);
+  const dayNum = d.getUTCDay() || 7;
+  d.setUTCDate(d.getUTCDate() + 4 - dayNum);
+  const yearStart = new Date(Date.UTC(d.getUTCFullYear(), 0, 1));
+  return Math.ceil(((d - yearStart) / 86400000 + 1) / 7);
+};
+
+export const getWeek = (date) => {
+  // +2 for tuesday to monday week offset
+  var firstday = new Date(date.setDate(date.getDate() - date.getDay() + 2));
+  var lastday = new Date(date.setDate(date.getDate() - date.getDay() + 6 + 2));
+  const s = `${firstday.getDate() + " " + months[firstday.getMonth()]}-${lastday.getDate() + " " + months[lastday.getMonth()]} ${date.getFullYear()}`
+  return s
+}
